refactor(frontend): rabbit-review
